+++ db/upgrade.php
@@ -113,6 +113,17 @@ function xmldb_theme_adaptable_upgrade($oldversion = 0) {
         upgrade_plugin_savepoint(true, 2024100502, 'theme', 'adaptable');
     }
 
+    if ($oldversion < 2024100503) {
+        $value = get_config('theme_adaptable', 'msgbadgecolor');
+        if (!empty($value)) {
+            set_config('notbadgebackgroundcolour', $value, 'theme_adaptable');
+            // Prevent replacement when upgrade has already happened in a version for an older Moodle!
+            unset_config('msgbadgecolor', 'theme_adaptable');
+        }
+
+        upgrade_plugin_savepoint(true, 2024100503, 'theme', 'adaptable');
+    }
+
     // Automatic 'Purge all caches'....
     purge_all_caches();